+++ AutomatedTesting/Gem/PythonTests/assetpipeline/TestSuite_Periodic.py
@@ -10,13 +10,16 @@ import pytest
 
 from ly_test_tools.o3de.editor_test import EditorTestSuite, EditorSingleTest
 
+
 @pytest.mark.SUITE_periodic
 @pytest.mark.parametrize("project", ["AutomatedTesting"])
 @pytest.mark.parametrize("launcher_platform", ['windows_editor'])
 class TestAutomation(EditorTestSuite):
 
+    @pytest.mark.skip(reason="https://github.com/o3de/o3de/issues/14514")
     class SubID_NoChange_MeshChanged(EditorSingleTest):
         from .fbx_tests import hydra_SubID_NoChange_MeshChanged as test_module
 
+    @pytest.mark.skip(reason="https://github.com/o3de/o3de/issues/14514")
     class SubID_WarningReported_AssetRemoved(EditorSingleTest):
         from .fbx_tests import hydra_SubID_WarningReported_AssetRemoved as test_module
