+++ AutomatedTesting/Gem/PythonTests/assetpipeline/asset_processor_tests/asset_processor_gui_tests_2.py
@@ -109,6 +109,7 @@ class TestsAssetProcessorGUI_WindowsAndMac(object):
 
         assert key_value.lower() == "true", f"The fast scan setting found was {key_value}"
 
+    @pytest.mark.skip(reason="https://github.com/o3de/o3de/issues/14514")
     @pytest.mark.test_case_id("C3635822")
     @pytest.mark.BAT
     @pytest.mark.assetpipeline
@@ -172,6 +173,7 @@ class TestsAssetProcessorGUI_AllPlatforms(object):
     Tests for Asset Processor GUI To Run on All Supported Host Platforms
     """
 
+    @pytest.mark.skip(reason="https://github.com/o3de/o3de/issues/14514")
     @pytest.mark.test_case_id("C1591337")
     @pytest.mark.BAT
     @pytest.mark.assetpipeline
@@ -299,6 +301,7 @@ class TestsAssetProcessorGUI_AllPlatforms(object):
         result, _ = asset_processor.gui_process()
         assert result, "AP GUI failed"
 
+    @pytest.mark.skip(reason="https://github.com/o3de/o3de/issues/14514")
     @pytest.mark.test_case_id("C24168802")
     @pytest.mark.BAT
     @pytest.mark.assetpipeline
